snake_draw

+++ src/grid/snake/point.rs
@@ -1,4 +1,4 @@
 pub struct Point {
-    pub x: u32,
-    pub y: u32
+    pub x: usize,
+    pub y: usize
 }
